requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
Pillow==9.5.0
psycopg==3.1.8
pycodestyle==2.10.0
PyJWT==2.3.0
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
gunicorn==20.1.0
whitenoise==6.4.0
